Microcontroller code for Texas Two Step, initial commit
The Texas Two Step is a replacement hand controller for the Synta
Dual-Axis motor kit for EQ-3 and EQ-5 type equatorial mounts, with
the following improvements:

* Faster setting rates (original HC is limited to 8x sidereal)
* Analog 2-axis joystick to control setting rates (instead of
  rate slider switch combined with directional pad)
* ST-4 support with independent guiding (ST-4) and setting
  (joystick) rates.
* Main unit attaches to mount; RA, DEC, ST-4 cables move with
  telescope, and a single cable connects the control unit to
  a handheld joystick unit.
* Integrated rechargeable battery (to replace 4x D-cell pack,
  and eliminate the additional cable for power)
* All LEDs are dim and red to preserve dark adaptation
* Additional tracking rates (solar, lunar)
* Smaller and lighter
* Lower power draw

The Texas Two Step is powered by an Atmel AVR microprocessor
(currently Atmega328 on an Adafruit Pro Trinket at 12Mhz and
3.3V) and the stepper motors are driven by dual TI DRV8834
low-voltage stepper driver ICs. The DRV8834 is selected to
allow a common power source for the logic and the motors.

This initial commit implements basic joystick control for the
setting rates. It contains known bugs and likely also contains
unknown bugs, and is missing some of the planned features,
including, but not limited to:

* The joystick is not calibrated, resulting in incorrect
  tracking/setting rates when the joystick is centered/moved.
* Microstepping is not yet implemented.
* ST-4 support is not implemented.
* Sidereal tracking only; lunar and solar not implemented.
* Directional switching not implemented; RA tracking currently
  moves in Southern Hemisphere direction on Synta EQ-3 with
  dual-axis motors.