Add potentiometer ranges to axis attributes
Store the minimum and maximum range of the potentiometer in the axis
info struct, to allow trimming the joystick. Rearrange the attributes
within the struct to keep permanent attributes organized separately
from runtime variables, and use designated initializers so that the
initial assignment is more self-documenting. The ranges for both axes
remain [0,1023] until the actual values can be measured.

Remove the microstep pin attribute until microstepping is actually
implemented.