Map joystick read based on length of array, not index of last element
Ensure more even distribution of the joystick rate ranges by using the
length of the array rather than the index of the last element as the
top value in the range to map() to. Constrain the mapped value to the
index of the last element.

Delete the per-axis tracking of the minimum and maximum potentiometer
values: these were measured to be 0 and 1023, respectively.