Reclaim unused microstep pins for use as enable
The DRV8834 drivers are not able to generate sufficient torque when
microstepping at the low current ceiling (~140 mA) they have been
configured to chop to. Rather than increase the current and therefore
the power consumption, just abandon the idea of microstepping, since
RA tracking at 1x sidereal rate is already in the neighborhood of 10Hz
anyway. Since the microstep control pins are now hardwired, they can
be reassigned for use as the enable pins.