Add preprocessor check for clock speed
The accuracy of micros() is limited by an integer division of 64 by the
clock rate in hertz divided by one million. Since the AdaFruit Pro Trinket
originally used to prototype this project runs at 12 MHz, this rounded the
value 64/12 down from 5 1/3 to 5, causing inaccuracy in the tracking rate.

The current prototype is a standalone Atmega328 on a breadboard, clocked
at 12 MHz to match the original 3.3V Pro Trinket prototype. Add a check at
preprocessing time to validate that the clock speed will allow micros() to
work accurately. The prototype will be reclocked to either 8MHz (if logic
voltage remains at 3.3V) or 16MHz (if logic voltage is raised to 5V).