Reassign stepper driver control pins
Pins 11 through 13 will soon be reassigned to the ST-4 port, to allow the
ST-4 port to serve double duty as an ISP port for reprogramming the MCU.
The RA axis is assigned an enable pin again, to support enforcing a delay
between power on and stepping.

This sketch is no longer compatible with the Adafruit Pro Trinket, due to
use of pins that are unavailable on that board. (The recent change to
validate the clock rate already made this sketch incompatible with the
3V version of the Pro Trinket, which is clocked at 12 MHz.)