Wait 1ms after enabling motor
The DRV8834 datasheet specifies a 1ms minimum time between waking up the
driver and stepping. The RA axis currently intermittently does not work,
but can be recovered by unplugging the jumper from the enable pin on the
driver board to Vcc and plugging it back in again; add a 1ms delay after
enabling the driver in case the missing delay is responsible for this
intermittent failure.