Add support for multiple setting rates
Assign an additional analog pin to read a switch (SPDT on-off-on, read
common pin which is tied to a pair of voltage dividing resistors) to
set the tracking rate. Since the number of analog reads is no longer
the same as the number of axes, refactor the joystick reading code to
be more generic analog reading code and decouple it from the axes.

The us_per_15_degrees values are based upon a sidereal day of 23h 56m
4.0916s and a sidereal month of 27d 7h 43m 11.5s, both figures obtained
from the Wikipedia articles for "Sidereal time" and "Lunar month",
respectively.