Add support for ST-4 guiding
Assign pins 10 through 13 for ST-4, and add code to debounce ST-4
input and adjust the rate based on the combination of ST-4 input and
the guiding rate set by the GUIDING_RATE potentiometer. Pins 10 through
13 are selected because pins 11 through 13 are used for SPI; pin ST-4
pin 3 will be connected to pin 10 when the controller is on, and to
RESET when the controller is off, to allow for reprogramming the
microcontroller over the ST-4 port.