Fully wire up DRV8834 chips
The selection of external supporting components is based in part on
the example application circuit in the DRV8834 datasheet, and in part
on the design of the Pololu DRV8834 breakout board.

The voltage divider and current sense resistor values listed in the
schematic should program a current limit of 117mA per phase, which is
in line with the current limit selected by the potentiometers in the
prototype. The voltage divider resistors will probably be installed in
a DIP socket to allow reconfiguring the current limit by replacing
resistors (slightly less convenient than a potentiometer, but also
slightly more permanent.)